+++ lib/src/widgets/loading_widget.dart
@@ -40,6 +40,8 @@ class LoadingWidget extends StatelessWidget {
             Text(
               message!,
               style: textStyle ?? defaultStyle,
+              maxLines: null,
+              overflow: TextOverflow.visible,
             ),
           ],
         ],